Fix issue with links getting cleared out after creating a Pivotal ticket
The success handler in the `tag` call was calling `$('a')`, which selects all links on the page, instead of `$(<a>)`, which creates a new DOM node. This causes a bunch of links on the page which had non-text children to have their contents cleared out.

Also, this updates the generated DOM nodes to match Custora customizations.